LO4 Linkedlist Integer JAVA Project
