+++ nbproject/project.properties
@@ -73,7 +73,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=it20c_ErickRubin
-main.class=LinkedListInt.Main
+main.class=topic_2_linkedlist_int.Main
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
